pyproject.toml
[project]
name = "mentat-bot"
version = "0.2.1"
description = "Discord bot designed to streamline credential management and access tracking during red team engagements."
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "asyncclick>=8.1.8",
    "discord-py>=2.4.0",
    "python-json-logger>=3.2.1",
    "rich>=13.9.4",
]

[project.scripts]
mentat = "mentat.cli:cli"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "mkautodocs>=0.3.4.3",
    "mkdocs-material>=9.6.3",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.6",
]
